Add files via upload
REINFORCEMENT LEARNING FOR CONNECT 4 IN A LINE GAMING BOT TRAINING.
The application is a Connect 4 with the purpose of creating 4 tokens in a row horizontally, vertically, or diagonally. You could do this by strategically placing tokens and blocking the other's attempts to win. Good luck!

Developing an intelligent game-playing bot for the Connect 4 in a Line game using Reinforcement Learning (RL) presents a challenging task. The objective is to design and train an RL agent capable of making optimal moves in Connect 4, a classic two-player board game. The bot should learn to play strategically, making decisions to win or prevent the opponent from winning, ultimately achieving a high level of gameplay proficiency.